implémentation des fonctionnalités Tags pour examen prépa WPWD2024

+++ src/main/resources/db/migration/V10_1__create_roles_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE roles (
-                       id INT(11) NOT NULL AUTO_INCREMENT,
+                       id BIGINT (11) NOT NULL AUTO_INCREMENT,
                        role VARCHAR(30) NOT NULL,
                        PRIMARY KEY (id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
