+++ src/main/resources/db/migration/V14_1__create_user_role_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE role_user (
-                           role_id INT(11) NOT NULL,
-                           user_id INT(11) NOT NULL,
+                           role_id BIGINT (11) NOT NULL,
+                           user_id BIGINT (11) NOT NULL,
                            PRIMARY KEY (role_id, user_id),
                            FOREIGN KEY (role_id) REFERENCES roles(id),
                            FOREIGN KEY (user_id) REFERENCES users(id)
